Use nginx.org yum repo instead of epel
EPEL wasn't carrying the range of versions of NGINX that the official NGINX was carrying, so I changed from EPEL to NGINX.

Also, I changed the ./configure to use the "with-compat" option.  At the least, it makes the configure command much simpler... and possibly we might be able to use the module without recompiling (not sure).